refactor components for redux
